Separate out JITting from interpreting
This patch separates out the JITter from the interpreter.  It adds a
"null" JIT which just errors out.

After this patch is applied it is possible to build on non-x86-64
platforms and run the interpreter without worrying about the JIT.

This is achieved by adding a translate function-pointer member to
the ubpf_vm struct. ubpf_create() then sets the pointer up to the
correct JIT.  Finally ubpf_translate() now delegates to that function
pointer.

To enable code reuse some common code from ubpf_jit_x86_64.c is moved
into a new file ubpf_jit.c.

Also what was ubpf_translate() in ubpf_jit_x86_64.c is renmaed
ubpf_translate_x86_64.c.

The build process assumes that all backends can be compiled on all
platforms.  It then relies on the linker to not link in unused backends.

The code also assumes that only the JIT backend targeting the platform
the code is run on will is needed.